Static background scaling issue with multi-monitor setup.
If you have multiple monitors and turned on "static background" plus "hide scaling worksplaces" you will experience that one of the monitors is unplugged and then plugged back in that the background will not fill correctly the desktop when you bring up the workspaces. To overcome this issue this patch uses some code brought from cosmic-workspaces which listens to monitor changes and updates the static backgrounds accordingly.